Move the endpoint parser into its own module.
