+++ src/twisted/plugins/txacme_endpoint.py
@@ -1,5 +1,5 @@
 from txacme.client import LETSENCRYPT_DIRECTORY, LETSENCRYPT_STAGING_DIRECTORY
-from txacme.endpoint import _AcmeParser
+from txacme._endpoint_parser import _AcmeParser
 
 
 le_parser = _AcmeParser('le', LETSENCRYPT_DIRECTORY)
